Make the methods smaller
Make the methods smaller. Most of this could be accomplished via
automated refactoring. But after extracting some methods, some other
opportunities for code cleanup/simplification became clear. A couple of
observations: 1. A lot of duplication was eliminated (e.g. the string
constants). 2. Creating more methods actually ended up having the
side-effect of shortening the code in the case of game #2. 3. While
working on this, I found myself kind of annoyed by the use of unnamed
integers to represent the scores, so perhaps that is the issue that I
should take on next.